+++ template-parts/section-no-results.php
@@ -12,7 +12,10 @@
 
 	<div class="page-content row-with-vspace">
 		<?php if (is_home() && current_user_can('publish_posts')) { ?> 
-			<p><?php printf(__('Ready to publish your first post? <a href="%1$s">Get started here</a>.', 'bootstrap-basic4'), esc_url(admin_url('post-new.php'))); ?></p>
+			<p><?php 
+                        /* translators: %1$s: URL to add new post. */
+                        printf(__('Ready to publish your first post? <a href="%1$s">Get started here</a>.', 'bootstrap-basic4'), esc_url(admin_url('post-new.php'))); 
+                        ?></p>
 		<?php } elseif (is_search()) { ?> 
 			<?php get_template_part('template-parts/partial', 'search-form'); ?> 
 			<p><?php _e('Sorry, but nothing matched your search terms. Please try again with some different keywords.', 'bootstrap-basic4'); ?></p>